Rename variable RUST_SUFFX to XBUILT
The variable contains "-x" if the build and target platforms are
not the same and "" (empty string) otherwise, and gets appended to
snapshot, installation tarball, and log filenames to denote whether
they have cross-built or natively-built content. However, this is
not really rust specific, may be useful in the future once other
components are also supported by the build system, thus renaming.